Change the data structure used for interval lists
Switches from a linked list to a flat array of doubles. Allows all
intersect/"unite" operations to be defined as simple wrappers for a
single "merge" function that can do both union and intersection.

The program seems to interpret the intervals as [a;b] and not [a;b) - so
that is what I have implemented of course.